+++ engine/framework/component/Renderable.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "framework/component/Component.h"
+namespace ChikaEngine::Framework
+{
+
+    class Renderable : public Component
+    {
+    };
+} // namespace ChikaEngine::Framework