+++ engine/physics/include/PhysicsDescs.h
@@ -1,4 +1,8 @@
 #pragma once
+#include "framework/gameobject/GameObject.h"
+#include "math/quaternion.h"
+#include "math/vector3.h"
+
 #include <cstdint>
 namespace ChikaEngine::Physics
 {
@@ -13,6 +17,7 @@ namespace ChikaEngine::Physics
     // 提供默认的物理参数
     struct PhysicsInitDesc
     {
+        Math::Vector3 gravity = Math::Vector3(0.0, 9.8, 0.0);
     };
 
     struct PhysicsSystemDesc
@@ -24,12 +29,50 @@ namespace ChikaEngine::Physics
         PhysicsInitDesc initDesc;
     };
 
+    // 用于记录物理计算后的位置
+    struct PhysicsTransform
+    {
+        Math::Vector3 pos;
+        Math::Quaternion rot;
+    };
+
+    enum class RigidbodyShapes
+    {
+        Box,
+        Sphare
+    };
     struct RigidbodyCreateDesc
     {
+        RigidbodyShapes shape = RigidbodyShapes::Box;
+        PhysicsTransform transform;
+        Math::Vector3 halfExtents{0.5f, 0.5f, 0.5f};
+        float radius = 0.5f;
+        float mass = 1.0f;
+        bool isKinematic = false;
+        Framework::GameObjectID ownerId = 0; // 用于查找 go
     };
 
     // 用于Collider回调
     struct CollisionEvent
     {
+        PhysicsBodyHandle selfRigidbodyHandle;
+        PhysicsBodyHandle oherRigidbodyHandle;
+        Math::Vector3 contactPoint;
+        Math::Vector3 contactNormal;
+        float impulse;
+    };
+
+    // 处理修改速度事件的数据
+    struct VelocityCommand
+    {
+        PhysicsBodyHandle handle;
+        Math::Vector3 v;
+    };
+
+    // 处理冲量的数据
+    struct ImpulseCommand
+    {
+        PhysicsBodyHandle handle;
+        Math::Vector3 impulse;
     };
 } // namespace ChikaEngine::Physics